[Delete] CalendarBottomCard 삭제, BottomCard 그대로 사용
